+++ tests/unit/test_supermetrics.py
@@ -1,4 +1,5 @@
 import pytest
+
 from viadot.sources import Supermetrics
 
 RESPONSE_PIVOTED = {