+++ viadot/flows/duckdb_to_sql_server.py
@@ -46,6 +46,8 @@ class DuckDBToSQLServer(Flow):
         dtypes: Dict[str, Any] = None,
         if_exists: Literal["fail", "replace", "append", "delete"] = "replace",
         sql_server_credentials: dict = None,
+        on_bcp_error: Literal["skip", "fail"] = "fail",
+        bcp_error_log_path="./log_file.log",
         tags: List[str] = ["load"],
         *args: List[any],
         **kwargs: Dict[str, Any],
@@ -67,6 +69,8 @@ class DuckDBToSQLServer(Flow):
             we infer them from the DataFrame. Defaults to None.
             if_exists (Literal, optional): What to do if the table exists. Defaults to "replace".
             sql_server_credentials (dict, optional): The credentials to use for connecting with SQL Server.
+            on_bcp_error (Literal["skip", "fail"], optional): What to do if error occurs. Defaults to "fail".
+            bcp_error_log_path (string, optional): Full path of an error file. Defaults to "./log_file.log".
             tags (List[str], optional): Flow tags to use, eg. to control flow concurrency. Defaults to ["load"].
         """
 
@@ -86,6 +90,8 @@ class DuckDBToSQLServer(Flow):
         self.dtypes = dtypes
         self.if_exists = self._map_if_exists(if_exists)
         self.sql_server_credentials = sql_server_credentials
+        self.on_bcp_error = on_bcp_error
+        self.bcp_error_log_path = bcp_error_log_path
 
         # Global
         self.tags = tags
@@ -137,6 +143,8 @@ class DuckDBToSQLServer(Flow):
             schema=self.sql_server_schema,
             table=self.sql_server_table,
             credentials=self.sql_server_credentials,
+            on_error=self.on_bcp_error,
+            error_log_file_path=self.bcp_error_log_path,
             flow=self,
         )
         cleanup_csv_task.bind(path=self.local_file_path, flow=self)