+++ viadot/flows/epicor_to_duckdb.py
@@ -1,8 +1,9 @@
-from prefect import Flow
 from typing import Any, Dict, List, Literal
 
-from ..tasks import EpicorOrdersToDF, DuckDBCreateTableFromParquet
-from ..task_utils import df_to_parquet, add_ingestion_metadata_task
+from prefect import Flow
+
+from ..task_utils import add_ingestion_metadata_task, cast_df_to_str, df_to_parquet
+from ..tasks import DuckDBCreateTableFromParquet, EpicorOrdersToDF
 
 
 class EpicorOrdersToDuckDB(Flow):
@@ -19,6 +20,7 @@ class EpicorOrdersToDuckDB(Flow):
         duckdb_table: str = None,
         duckdb_schema: str = None,
         if_exists: Literal["fail", "replace", "append", "skip", "delete"] = "fail",
+        if_empty: Literal["warn", "skip", "fail"] = "skip",
         duckdb_credentials: dict = None,
         *args: List[any],
         **kwargs: Dict[str, Any],
@@ -38,6 +40,7 @@ class EpicorOrdersToDuckDB(Flow):
             duckdb_table (str, optional): Destination table in DuckDB. Defaults to None.
             duckdb_schema (str, optional): Destination schema in DuckDB. Defaults to None.
             if_exists (Literal, optional):  What to do if the table already exists. Defaults to "fail".
+            if_empty (Literal, optional): What to do if Parquet file is empty. Defaults to "skip".
             duckdb_credentials (dict, optional): Credentials for the DuckDB connection. Defaults to None.
         """
         self.base_url = base_url
@@ -50,6 +53,7 @@ class EpicorOrdersToDuckDB(Flow):
         self.duckdb_table = duckdb_table
         self.duckdb_schema = duckdb_schema
         self.if_exists = if_exists
+        self.if_empty = if_empty
         self.duckdb_credentials = duckdb_credentials
 
         super().__init__(*args, name=name, **kwargs)
@@ -72,8 +76,8 @@ class EpicorOrdersToDuckDB(Flow):
             end_date_field=self.end_date_field,
             start_date_field=self.start_date_field,
         )
-        df_with_metadata = add_ingestion_metadata_task.bind(df, flow=self)
-
+        df_mapped = cast_df_to_str.bind(df, flow=self)
+        df_with_metadata = add_ingestion_metadata_task.bind(df_mapped, flow=self)
         parquet = df_to_parquet.bind(
             df=df_with_metadata,
             path=self.local_file_path,
@@ -85,6 +89,7 @@ class EpicorOrdersToDuckDB(Flow):
             schema=self.duckdb_schema,
             table=self.duckdb_table,
             if_exists=self.if_exists,
+            if_empty=self.if_empty,
             flow=self,
         )
         create_duckdb_table.set_upstream(parquet, flow=self)