+++ viadot/sources/bigquery.py
@@ -1,6 +1,7 @@
 from typing import List
-import pandas_gbq
+
 import pandas as pd
+import pandas_gbq
 from google.oauth2 import service_account
 
 from ..config import local_config