+++ viadot/tasks/__init__.py
@@ -1,11 +1,12 @@
+from .aselite import ASELiteToDF
 from .azure_blob_storage import BlobFromCSV
 from .azure_data_lake import (
     AzureDataLakeCopy,
     AzureDataLakeDownload,
     AzureDataLakeList,
+    AzureDataLakeRemove,
     AzureDataLakeToDF,
     AzureDataLakeUpload,
-    AzureDataLakeRemove,
 )
 from .azure_key_vault import (
     AzureKeyVaultSecret,
@@ -16,23 +17,22 @@ from .azure_sql import (
     AzureSQLBulkInsert,
     AzureSQLCreateTable,
     AzureSQLDBQuery,
-    CreateTableFromBlob,
     AzureSQLToDF,
-    CheckColumnOrder,
     AzureSQLUpsert,
+    CheckColumnOrder,
+    CreateTableFromBlob,
 )
 from .bcp import BCPTask
+from .bigquery import BigQueryToDF
+from .cloud_for_customers import C4CReportToDF, C4CToDF
 from .github import DownloadGitHubFile
 from .great_expectations import RunGreatExpectationsValidation
-from .sqlite import SQLiteInsert, SQLiteSQLtoDF, SQLiteQuery
-from .supermetrics import SupermetricsToCSV, SupermetricsToDF
-from .sharepoint import SharepointToDF
-from .cloud_for_customers import C4CReportToDF, C4CToDF
-from .prefect_date_range import GetFlowNewDateRange
-from .aselite import ASELiteToDF
-from .bigquery import BigQueryToDF
-from .salesforce import SalesforceUpsert, SalesforceBulkUpsert
 from .outlook import OutlookToDF
+from .prefect_date_range import GetFlowNewDateRange
+from .salesforce import SalesforceBulkUpsert, SalesforceToDF, SalesforceUpsert
+from .sharepoint import SharepointToDF
+from .sqlite import SQLiteInsert, SQLiteQuery, SQLiteSQLtoDF
+from .supermetrics import SupermetricsToCSV, SupermetricsToDF
 
 try:
     from .sap_rfc import SAPRFCToDF
@@ -41,4 +41,5 @@ except ImportError:
 
 from .duckdb import DuckDBCreateTableFromParquet, DuckDBQuery, DuckDBToDF
 from .sql_server import SQLServerCreateTable, SQLServerToDF, SQLServerQuery
+
 from .epicor import EpicorOrdersToDF