+++ CHANGELOG.md
@@ -7,10 +7,60 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 
+- Added handling empty DF in `set_new_kv()` task
+- Added `update_kv` and `filter_column` params to `SAPRFCToADLS` and `SAPToDuckDB` flows and added `set_new_kv()` task
+in `task_utils`
+
+### Changed
+- Changed default value of `on_error` parameter in `BCPTask` and `on_bcp_error` parameter in `ADLSToAzureSQL` 
+and `DuckDBToSQLServer` to `fail`. 
+## [0.4.6] - 2022-07-21
+### Added
+- Added `rfc_character_limit` parameter in `SAPRFCToDF` task, `SAPRFC` source, `SAPRFCToADLS` and `SAPToDuckDB` flows
+- Added `on_bcp_error` and `bcp_error_log_path` parameters in `BCPTask`
+- Added ability to process queries which result exceed SAP's character per low limit in `SAPRFC` source
+- Added new flow `PrefectLogs` for extracting all logs from Prefect with details
+- Added `PrefectLogs` flow
+
+### Changed
+- Changed `CheckColumnOrder` task and `ADLSToAzureSQL` flow to handle appending to non existing table
+- Changed tasks order in `EpicorOrdersToDuckDB`, `SAPToDuckDB` and `SQLServerToDuckDB` - casting 
+DF to string before adding metadata
+- Changed `add_ingestion_metadata_task()` to not to add metadata column when input DataFrame is empty
+- Changed `check_if_empty_file()` logic according to changes in `add_ingestion_metadata_task()`
+- Changed accepted values of `if_empty` parameter in `DuckDBCreateTableFromParquet`
+- Updated `.gitignore` to ignore files with `*.bak` extension and to ignore `credentials.json` in any directory
+- Changed logger messages in `AzureDataLakeRemove` task
+
+### Fixed
+- Fixed handling empty response in `SAPRFC` source
+- Fixed issue in `BCPTask` when log file couln't be opened.
+- Fixed log being printed too early in `Salesforce` source, which would sometimes cause a `KeyError`
+- `raise_on_error` now behaves correctly in `upsert()` when receiving incorrect return codes from Salesforce
+
+### Removed
+- Removed option to run multiple queries in `SAPRFCToADLS`
+
+
+## [0.4.5] - 2022-06-23
+### Added
+- Added `error_log_file_path` parameter in `BCPTask` that enables setting name of errors logs file 
+- Added `on_error` parameter in `BCPTask` that tells what to do if bcp error occurs. 
+- Added error log file and `on_bcp_error` parameter in `ADLSToAzureSQL`
+- Added handling POST requests in `handle_api_response()` add added it to `Epicor` source.
+- Added `SalesforceToDF` task
+- Added `SalesforceToADLS` flow
+- Added `overwrite_adls` option to `BigQueryToADLS` and `SharepointToADLS`
+- Added `cast_df_to_str` task in `utils.py` and added this to `EpicorToDuckDB`, `SAPToDuckDB`, `SQLServerToDuckDB`
+- Added `if_empty` parameter in `DuckDBCreateTableFromParquet` task and in `EpicorToDuckDB`, `SAPToDuckDB`,
+`SQLServerToDuckDB` flows to check if output Parquet is empty and handle it properly.
+- Added `check_if_empty_file()` and `handle_if_empty_file()` in `utils.py`
+
+>>>>>>> 21c833ac5ae24e1806481abbc341e780e5ffc751
+
 - Added new flow - `SQLServerTransform` and new task `SQLServerQuery` to run queries on SQLServer
 ## [0.4.4] - 2022-06-09
 ### Added
-
 - Added new connector - Outlook. Created `Outlook` source, `OutlookToDF` task and `OutlookToADLS` flow.
 - Added new connector - Epicor. Created `Epicor` source, `EpicorToDF` task and `EpicorToDuckDB` flow.
 - Enabled Databricks Connect in the image. To enable, [follow this guide](./README.md#executing-spark-jobs)